(Milestone): Completed the account-details feature and transfer feature

+++ frontend/web/src/app/pages/main/account_details/account-details.ts
@@ -8,6 +8,7 @@ import { AccountsService } from '@api/accounts.service';
 import { DrawerService } from '@infrastructure/services';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NoData } from '@structural/main/no-data/no-data';
+import { AccountDetailsForm } from './account-details.form';
 import { Component, computed, inject, signal } from '@angular/core';
 import { TransactionsComponent } from './transactions/transactions';
 import { FetchError } from '@structural/main/fetch-error/fetch-error';
@@ -22,6 +23,7 @@ import { LucideAngularModule, Wallet, SquarePen, Trash2 } from 'lucide-angular';
     FetchError,
     Breadcrumbs,
     DetailsComponent,
+    AccountDetailsForm,
     LucideAngularModule,
     TransactionsComponent,
   ],
@@ -35,6 +37,7 @@ export class AccountDetails {
 
   // SIGNALS
   protected readonly deleteClickCount = signal<1 | 2>(1);
+  protected readonly isEditFormOpen = signal<boolean>(false);
   protected readonly isDeletingAccount = signal<boolean>(false);
 
   // SERVICES
@@ -76,19 +79,29 @@ export class AccountDetails {
   });
 
   // METHODS
-  private reloadResources = () => this.accountsService.getUserAccounts().reload();
+  private reloadResources = () => {
+    this.accountWithTransactions.reload();
+    this.accountsService.getUserAccounts().reload();
+  };
 
   protected handleOnEditClick() {
-    console.log('Edit clicked');
+    this.isEditFormOpen.set(true);
+  }
+
+  protected handleEditFormClose(cause: 'submit' | 'icon' | 'overlay') {
+    if (cause === 'submit') this.reloadResources();
+    this.isEditFormOpen.set(false);
   }
 
   protected handleOnDeleteAccountClick(accountId: string) {
+    if (!accountId || accountId === '') return;
+
     if (this.deleteClickCount() === 1) {
       this.deleteClickCount.set(2);
       this.toastService.show({
         variant: 'warning',
         title: 'Are you sure?',
-        details: 'Deleting an account will remove all transactions associated with it.',
+        details: 'Deleting an account will also delete all transactions associated with it.',
       });
       return;
     }
